requirements.txt
python-kraken-sdk==1.6.2
python-dotenv==1.0.0
requests==2.31.0
